+++ NEWS.md
@@ -4,6 +4,7 @@
 Released 2017-mm-dd
 
 Announcements:
+ - Upgrades windshaft to [3.2.1](https://github.com/CartoDB/windshaft/releases/tag/3.2.1).
  - Add support to retrieve info about layer stats in map instantiation.
  - Upgrades camshaft to [0.55.1](https://github.com/CartoDB/camshaft/releases/tag/0.55.1).
  - Remove promise polyfill from turbo-carto adapter